Add most of what makes D-Bus tick
Object and proxy support, mapping for all relevant types, methods,
properties, emitting and subscribing to signals.

Added license information headers in all source code, expanded the
Meson build script to compile a separate library and two examples
(a "service" and proxy that queries it).

Added a skeleton README.md, to be expanded once the work
stabilizes.